move build-in jsch plugin to a core plugins
