+++ core/build.gradle
@@ -79,14 +79,6 @@ dependencies {
         'commons-lang:commons-lang:2.6',
         'jaxen:jaxen:1.2.0',
         'commons-cli:commons-cli:1.0',
-        'org.apache.ant:ant-jsch:1.9.16',
-        'com.jcraft:jsch:0.1.55',
-        'com.jcraft:jsch.agentproxy.jsch:0.0.9',
-        'com.jcraft:jsch.agentproxy.sshagent:0.0.9',
-        'com.jcraft:jsch.agentproxy.usocket-jna:0.0.9',
-        'com.jcraft:jsch.agentproxy.usocket-nc:0.0.9',
-        'com.jcraft:jsch.agentproxy.connector-factory:0.0.9',
-        'com.jcraft:jsch.agentproxy.core:0.0.9',
         "com.fasterxml.jackson.core:jackson-databind:${jacksonDatabindVersion}",
         "org.yaml:snakeyaml:${snakeyamlVersion}",
         'com.squareup.retrofit2:retrofit:2.9.0',
