+++ plugins/jsch-plugin/src/main/java/org/rundeck/plugins/jsch/net/ExtSSHExec.java
@@ -23,9 +23,9 @@
  * @version $Revision$
  */
 
-package com.dtolabs.rundeck.core.tasks.net;
+package org.rundeck.plugins.jsch.net;
 
-import com.dtolabs.rundeck.core.utils.SSHAgentProcess;
+import org.rundeck.plugins.jsch.util.SSHAgentProcess;
 import com.dtolabs.rundeck.plugins.PluginLogger;
 import com.jcraft.jsch.*;
 import org.apache.tools.ant.BuildException;
