+++ plugins/jsch-plugin/src/test/groovy/org/rundeck/plugins/jsch/JschSecretBundleUtilSpec.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.dtolabs.rundeck.core.execution.impl.jsch
+package org.rundeck.plugins.jsch
 
 import com.dtolabs.rundeck.core.common.Framework
 import com.dtolabs.rundeck.core.common.IRundeckProject
