+++ plugins/jsch-plugin/src/test/groovy/org/rundeck/plugins/jsch/SudoResponderSpec.groovy
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package com.dtolabs.rundeck.core.execution.impl.jsch
+package org.rundeck.plugins.jsch
 
-import com.dtolabs.rundeck.core.common.Framework
+
 import com.dtolabs.rundeck.core.common.NodeEntryImpl
 import com.dtolabs.rundeck.core.execution.ExecutionContext
 import com.dtolabs.rundeck.core.execution.utils.BasicSource
 import com.dtolabs.rundeck.core.execution.utils.PasswordSource
-import com.dtolabs.rundeck.core.tools.AbstractBaseTest
+import org.rundeck.plugins.jsch.util.JschTestUtil
 import spock.lang.Specification
 
 /**
@@ -34,7 +34,7 @@ class SudoResponderSpec extends Specification {
     def "sudo not enabled"(String password, byte[] expected, boolean enabled) {
         setup:
         def node = new NodeEntryImpl("test")
-        def fwk = AbstractBaseTest.createTestFramework()
+        def fwk = JschTestUtil.createTestFramework()
         fwk.getFrameworkProjectMgr().createFrameworkProject("SudoResponderTest")
         def pwdsource = Mock(PasswordSource) {
             getPassword() >> password.bytes
@@ -57,7 +57,7 @@ class SudoResponderSpec extends Specification {
         setup:
         def node = new NodeEntryImpl("test")
         node.getAttributes().put("sudo-command-enabled", "true")
-        def fwk = AbstractBaseTest.createTestFramework()
+        def fwk = JschTestUtil.createTestFramework()
         fwk.getFrameworkProjectMgr().createFrameworkProject("SudoResponderTest")
         def pwdsource = new BasicSource(password.bytes)
         def resp = SudoResponder.create(node, fwk, Mock(ExecutionContext) {
